+++ RuntimeEditor/EditorIcon.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using UnityEngine;
 using Hananoki;
-namespace HananokiEditor {
+namespace Hananoki {
 	public static class EditorIcon  {
 		public static Texture2D Error => Icon.Get( "console.erroricon.sml" );
 		public static Texture2D Warning => Icon.Get( "console.warnicon.sml" );