+++ RuntimeEditor/GUIDUtils.cs
@@ -3,7 +3,7 @@
 using UnityEditor;
 using System.IO;
 
-namespace HananokiEditor {
+namespace Hananoki {
 	public static class GUIDUtils {
 		public static T LoadAssetAtGUID<T>( string guid ) where T : UnityEngine.Object {
 			var path = UnityEditor.AssetDatabase.GUIDToAssetPath( guid );
@@ -11,11 +11,7 @@ namespace HananokiEditor {
 			return b;
 		}
 
-		public static string GetAssetGUID( UnityEngine.Object obj ) {
-			var a = AssetDatabase.GetAssetPath( obj );
-			var b = AssetDatabase.AssetPathToGUID( a );
-			return b;
-		}
+		
 
 		public static string GetAssetPath( string guid ) {
 			return AssetDatabase.GUIDToAssetPath( guid );
@@ -24,6 +20,12 @@ namespace HananokiEditor {
 
 		public static string ToGUID( this string path ) {
 			var b = AssetDatabase.AssetPathToGUID( path );
+			return b;
+		}
+
+		public static string ToGUID( UnityEngine.Object obj ) {
+			var a = AssetDatabase.GetAssetPath( obj );
+			var b = AssetDatabase.AssetPathToGUID( a );
 			return b;
 		}